autogen-telegram/requirements.txt
pyautogen==0.2.35
python-telegram-bot==21.7
httpx[socks]
